backend/pyproject.toml
[project]
name = "sfms"
version = "0.1.0"
description = "Sports Facility Management System - Backend API"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [{name = "Girish Hiremath"}]

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "httpx>=0.27",
    "asyncpg>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.9",
    "structlog>=24.1",
    "slowapi>=0.1.9",
]

[dependency-groups]
test = ["pytest>=8.0", "pytest-cov>=5.0", "pytest-asyncio>=0.24", "httpx"]
quality = ["ruff>=0.8", "mypy>=1.13", "bandit[toml]>=1.7"]
dev = [{include-group = "test"}, {include-group = "quality"}]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = ["--cov=src", "--cov-report=term-missing", "--verbose"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
fail_under = 60
show_missing = true

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["sfms"]

[tool.mypy]
python_version = "3.12"
mypy_path = "src"
strict = true
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests"]
